Replace deprecated function calls
Some replacements are only available in Qt 5.10 above, so keep the
old versions behind a preprocessor conditional until minimum supported
Qt version is raised.

+++ src/libshared/util/passwordhash.cpp
@@ -75,7 +75,7 @@ bool isValidHash(const QByteArray &hash)
 
 	} else if(hash.startsWith("sodium;")) {
 #ifdef HAVE_LIBSODIUM
-		return hash.length() > 7 && hash.length() < 7+crypto_pwhash_STRBYTES;
+		return hash.length() > 7 && (unsigned int)hash.length() < 7+crypto_pwhash_STRBYTES;
 #else
 		qWarning("Libsodium needed to support Argon2 hashes!");
 		return false;
